cli/requirements.txt
# requirements.txt

werkzeug
openai
python-dotenv
tqdm
tiktoken
nltk
moviepy
pydub